dom/tests/mochitest/mouselock/test_DOMtree.html

<!DOCTYPE HTML>
<html>
    <!--
    https://bugzilla.mozilla.org/show_bug.cgi?id=633602
    -->
    <head>
        <title>Test for Bug 633602</title>
        <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
        <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
        <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    </head>
    <body>
        <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
        <div id="content">
        </div>
        <pre id="test">
            <script type="text/javascript">
                /** Test for Bug 633602 **/
                SimpleTest.waitForExplicitFinish()
                SimpleTest.waitForFocus(function() {
                    /*------------test start----------*/
                    SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
                    
                    var pointer = navigator.pointer;
                    var canvas = document.createElement('canvas');
                    canvas.id = "testCanvas";
                    canvas.width = 250;
                    canvas.height = 250;
                    canvas.className = "test";
                    
                    pointer.lock(canvas);
                    is(pointer.islocked(), false, "Mouse should NOT be locked when DOM element is not in the tree.");
                    
                    var bodyTag = document.getElementsByTagName('body')[0];
                    bodyTag.appendChild(canvas);
                    canvas.mozRequestFullScreen();
                    
                    document.addEventListener("mozfullscreenchange", function (e) {
                        if (document.mozFullScreen) {
                            // Mouse should be lock if the element is in fullscreen mode
                            pointer.lock(canvas);
                            is(pointer.islocked(), true, "Mouse should be locked when DOM element is in the tree."); //if still fail, throw to inside the mozfullscreenchange event
                            pointer.unlock();
                            document.mozCancelFullScreen();
                        }
                        else {
                            bodyTag.removeChild(canvas);
                            is(pointer.islocked(), false, "Mouse should NOT be locked when DOM element is removed from the tree.");
                            /*------------test end----------*/
                            SimpleTest.finish();
                        }
                    }, false);
                });
            </script>
        </pre>
    </body>
</html>
